requirements.txt
absl-py==2.1.0
accelerate==0.30.1
albumentations==1.4.4
decord==0.6.0
diffusers==0.30.2
docstring_parser==0.16
easydict==1.13
einops==0.8.0
# flash-attn==2.5.8 # pip install https://ghfast.top/github.com/Dao-AILab/flash-attention/releases/download/v2.7.3/flash_attn-2.7.3+cu12torch2.3cxx11abiFALSE-cp311-cp311-linux_x86_64.whl
fvcore==0.1.5.post20221221
gsutil==5.30
h5py==3.11.0
huggingface-hub==0.24.6
hydra-core==1.3.2
imageio==2.34.1
imageio-ffmpeg==0.3.0
immutabledict==4.2.0
importlib_metadata==8.4.0
importlib_resources==6.4.0
iopath==0.1.9
Jinja2==3.1.2
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
lazy_loader==0.4
libclang==18.1.1
lightning==2.2.5
lightning-utilities==0.11.6
Markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.9.0
ml-dtypes==0.4.0
multidict==6.1.0
multiprocess==0.70.16
ninja==1.11.1.1
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
pandas==2.2.2
pathspec==0.12.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-xlib==0.33
pytorch-lightning==2.3.3
pytorch-seed==0.2.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.1
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
tensorboard==2.17.1
tensorflow==2.17.0
timm==1.0.3
tokenizers==0.21.0
# torch==2.3.0 # pip install torch==2.3.0 torchvision==0.18.0 torchaudio==2.3.0 --index-url https://download.pytorch.org/whl/cu121
# torchaudio==2.3.1
# torchelastic==0.2.2
# torchmetrics==1.4.2
# torchvision==0.18.1
tqdm
transformers==4.48.0
triton==2.3.0
urllib3==2.0.5
wandb==0.17.5
xformers==0.0.26.post1
